Folder structure
- Moved <ActionForm> into components folder
- Created Resolutions folder and moved related files
