+++ src/views/GovernanceReviewTeam/Actions/ManageLcid/RetireLcid.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import ActionForm, { SystemIntakeActionFields } from '../ActionForm';
+import ActionForm, { SystemIntakeActionFields } from '../components/ActionForm';
 
 interface RetireLcidFields extends SystemIntakeActionFields {}
