+++ src/views/GovernanceReviewTeam/Actions/RequestEdits.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { Trans, useTranslation } from 'react-i18next';
 
-import ActionForm, { SystemIntakeActionFields } from './ActionForm';
+import ActionForm, { SystemIntakeActionFields } from './components/ActionForm';
 
 interface RequestEditsFields extends SystemIntakeActionFields {}
